Refactored sdk package name
