+++ AppLaunchClientAndroid/.idea/modules.xml
@@ -3,7 +3,10 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/AppLaunchClientAndroid.iml" filepath="$PROJECT_DIR$/AppLaunchClientAndroid.iml" />
+      <module fileurl="file://$PROJECT_DIR$/AppLaunchClientAndroid.iml" filepath="$PROJECT_DIR$/AppLaunchClientAndroid.iml" />
       <module fileurl="file://$PROJECT_DIR$/applaunch-client-android/applaunch-client-android.iml" filepath="$PROJECT_DIR$/applaunch-client-android/applaunch-client-android.iml" />
+      <module fileurl="file://$PROJECT_DIR$/applaunch-client-android/applaunch-client-android.iml" filepath="$PROJECT_DIR$/applaunch-client-android/applaunch-client-android.iml" />
+      <module fileurl="file://$PROJECT_DIR$/applaunchsample/applaunchsample.iml" filepath="$PROJECT_DIR$/applaunchsample/applaunchsample.iml" />
       <module fileurl="file://$PROJECT_DIR$/applaunchsample/applaunchsample.iml" filepath="$PROJECT_DIR$/applaunchsample/applaunchsample.iml" />
     </modules>
   </component>
