+++ AppLaunchClientAndroid/applaunch-client-android/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.applaunch">
+    package="com.ibm.mobile.applaunch.android">
 
     <application
         android:allowBackup="true"
