--- AppLaunchClientAndroid/applaunch-client-android/src/main/java/com/applaunch/AppLaunchResponseListener.java
@@ -1,12 +1,0 @@
-package com.applaunch;
-
-
-
-/**
- * Created by norton on 7/25/17.
- */
-
-public interface AppLaunchResponseListener {
-    public void onSuccess(AppLaunchResponse appLaunchResponse);
-    public void onFailure(AppLaunchFailResponse appLaunchFailResponse);
-}
