--- AppLaunchClientAndroid/applaunch-client-android/src/main/java/com/applaunch/actions/Actions.java
@@ -1,21 +1,0 @@
-package com.applaunch.actions;
-
-/**
- * Created by norton on 10/26/17.
- */
-
-public class Actions {
-
-    public static Actions thisInstance =null;
-
-    private Actions() {
-    }
-
-    public static Actions getInstance(){
-        if(thisInstance==null){
-            thisInstance = new Actions();
-        }
-        return  thisInstance;
-    }
-
-}
