--- AppLaunchClientAndroid/applaunch-client-android/src/main/java/com/applaunch/api/AppLaunchActions.java
@@ -1,11 +1,0 @@
-package com.applaunch.api;
-
-/**
- * Created by norton on 9/22/17.
- */
-
-public interface AppLaunchActions {
-
-    public void onFeaturesReceived(String features);
-//    public void onMessagesReceived();
-}
