--- AppLaunchClientAndroid/applaunch-client-android/src/main/java/com/applaunch/api/AppLaunchException.java
@@ -1,15 +1,0 @@
-package com.applaunch.api;
-
-/**
- * Created by norton on 9/22/17.
- */
-
-public class AppLaunchException extends Exception {
-
-    public AppLaunchException() {
-    }
-
-    public AppLaunchException(String message) {
-        super(message);
-    }
-}
