--- AppLaunchClientAndroid/applaunch-client-android/src/main/java/com/applaunch/api/ReturnValue.java
@@ -1,22 +1,0 @@
-package com.applaunch.api;
-
-/**
- * Created by norton on 8/4/17.
- */
-
-class ReturnValue {
-
-    private boolean returnValue;
-
-    protected ReturnValue() {
-        super();
-    }
-
-    protected boolean returnValue() {
-        return returnValue;
-    }
-
-    protected void setReturnValue(boolean returnValue) {
-        this.returnValue = returnValue;
-    }
-}
