--- AppLaunchClientAndroid/applaunch-client-android/src/main/java/com/applaunch/common/AppLaunchConstants.java
@@ -1,34 +1,0 @@
-package com.applaunch.common;
-
-
-/**
- * Created by norton on 9/22/17.
- */
-
-public class AppLaunchConstants {
-
-    public static final String VERSION = "version";
-    public static final String ENGAGE_CLIENT_PROPS_NAME = "engageclient.properties";
-    public static final String ENGAGE_SERVER_HOST = "engageServerHost";
-    public static final String ENGAGE_SERVER_PROTOCOL = "engageServerProtocol";
-    public static final String ENGAGE_SERVER_PORT = "engageServerPort";
-    public static final String ENGAGE_SERVER_CONTEXT = "engageServerContext";
-    public static final String ANALYZER_SERVER_HOST = "analyzerServerHost";
-    public static final String ANALYZER_SERVER_PROTOCOL = "analyzerServerprotocol";
-    public static final String ANALYZER_SERVER_PORT = "analyzerServerport";
-    public static final String ANALYZER_SERVER_CONTEXT = "analyzerServerContext";
-
-    //URL
-    public static final String REGION_US_SOUTH_STAGING="mobileservices-staging.us-south.containers.mybluemix.net";
-    public static final String REGION_US_SOUTH_DEV="mobileservices-dev.us-south.containers.mybluemix.net";
-    public static final String REGION_US_SOUTH="mobileservices.us-south.containers.mybluemix.net";
-
-    //sharedpref keys
-    //public static final String FIRST_TIME_USER = "isFirstTime";
-    public static final String REG_RESPONSE = "regResponse";
-    public static final String ACTIONS_INVOKED="actionsInvoked";
-    public static final String ANALYZER_URL="analyzerUrl";
-    public static final String APP_USER="appUser";
-
-
-}
