+++ AppLaunchClientAndroid/applaunch-client-android/src/main/java/com/ibm/mobile/applaunch/android/AppLaunchResponseListener.java
@@ -1,0 +1,12 @@
+package com.ibm.mobile.applaunch.android;
+
+
+
+/**
+ * Created by norton on 7/25/17.
+ */
+
+public interface AppLaunchResponseListener {
+    public void onSuccess(AppLaunchResponse appLaunchResponse);
+    public void onFailure(AppLaunchFailResponse appLaunchFailResponse);
+}
