+++ AppLaunchClientAndroid/applaunch-client-android/src/main/java/com/ibm/mobile/applaunch/android/api/AppLaunchActions.java
@@ -1,0 +1,11 @@
+package com.ibm.mobile.applaunch.android.api;
+
+/**
+ * Created by norton on 9/22/17.
+ */
+
+public interface AppLaunchActions {
+
+    public void onFeaturesReceived(String features);
+//    public void onMessagesReceived();
+}
