+++ AppLaunchClientAndroid/applaunch-client-android/src/main/java/com/ibm/mobile/applaunch/android/api/AppLaunchException.java
@@ -1,0 +1,15 @@
+package com.ibm.mobile.applaunch.android.api;
+
+/**
+ * Created by norton on 9/22/17.
+ */
+
+public class AppLaunchException extends Exception {
+
+    public AppLaunchException() {
+    }
+
+    public AppLaunchException(String message) {
+        super(message);
+    }
+}
