+++ AppLaunchClientAndroid/applaunch-client-android/src/main/java/com/ibm/mobile/applaunch/android/api/MessageTypes.java
@@ -1,0 +1,13 @@
+package com.ibm.mobile.applaunch.android.api;
+
+/**
+ * Created by norton on 9/15/17.
+ */
+
+public interface MessageTypes {
+
+    public static String BANNER="banner";
+    public static String TOP_SLICE="topSlice";
+    public static String BOTTOM_PANEL="bottomPanel";
+
+}
